+++ test_problems/cxx_ex/kinetics_example2.cpp
@@ -36,7 +36,7 @@ int kinetics_example2(int job)
         // create a GRI30 object
         GRI30 gas;
         gas.setState_TPX(1001.0, OneAtm, "H2:2.0, O2:1.0, N2:4.0");
-        int kk = gas.nSpecies();
+        size_t kk = gas.nSpecies();
 
         // create a reactor
         Reactor r;
